requirements.txt
Django==5.2.1
djangorestframework==3.16.0
python-decouple==3.8
PyJWT==2.10.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0